Version 4.5.1
git-svn-id: http://plugins.svn.wordpress.org/wp-slimstat/trunk@1570855 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ browscap/browser.php
@@ -20,7 +20,9 @@ class slim_browser {
 			$error = self::update_browscap_database( false );
 			require_once( self::$browscap_autoload_path );
 
-			self::$browser = slimBrowscapConnector::get_browser_from_browscap( self::$browser );
+			if ( method_exists( 'slimBrowscapConnector', 'get_browser_from_browscap' ) ) {
+				self::$browser = slimBrowscapConnector::get_browser_from_browscap( self::$browser );
+			}
 		}
 	}
 
@@ -46,7 +48,7 @@ class slim_browser {
 	 */
 	public static function update_browscap_database( $_force_download = false ) {
 		if ( version_compare( PHP_VERSION, '5.5', '<' ) ) {
-			return array( 4, __( 'This library requires at least PHP 5.5. Please ask your service provider to upgrade your server accordingly.', 'wp-slimstat' ) );
+			return array( 1, __( 'This library requires at least PHP 5.5. Please ask your service provider to upgrade your server accordingly.', 'wp-slimstat' ) );
 		}
 
 		// Create the folder, if it doesn't exist
@@ -58,9 +60,9 @@ class slim_browser {
 		$local_version = 0;
 		$browscap_zip = wp_slimstat::$upload_dir . '/browscap-db.zip';
 
-		// Check for updates once a week ( 604800 seconds )
-		if ( false === $download_remote_file && false !== ( $file_stat = @stat( self::$browscap_autoload_path ) ) && ( date( 'U' ) - $file_stat[ 'mtime' ] > 604800 ) && false !== ( $handle = @fopen( self::$browscap_autoload_path, "rb" ) ) ) {
-			
+		// Check for updates once a week ( 604800 seconds ), if $_force_download is not true
+		if ( false === $_force_download && false !== ( $file_stat = @stat( self::$browscap_autoload_path ) ) && ( date( 'U' ) - $file_stat[ 'mtime' ] > 604800 ) && false !== ( $handle = @fopen( self::$browscap_autoload_path, "rb" ) ) ) {
+
 			// Find the version of the local data file
 			while ( ( $buffer = fgets( $handle, 4096 ) ) !== false ) {
 				if ( strpos( $buffer, 'source_version' ) !== false ) {
@@ -70,7 +72,7 @@ class slim_browser {
 			}
 
 			// Now check the version number on the server
-			$remote_version = file_get_contents( 'http://s3.amazonaws.com/browscap/autoload.txt' );
+			$remote_version = wp_remote_retrieve_body( wp_remote_get( 'http://s3.amazonaws.com/browscap/autoload.txt' ) );
 			if ( intval( $local_version ) != intval( $remote_version ) ) {
 				$download_remote_file = true;
 			}
@@ -84,17 +86,24 @@ class slim_browser {
 		// Download the most recent version of our pre-processed Browscap database
 		if ( $download_remote_file ) {
 			$response = wp_safe_remote_get( 'http://s3.amazonaws.com/browscap/browscap-db.zip', array( 'timeout' => 300, 'stream' => true, 'filename' => $browscap_zip ) );
-			if ( is_wp_error( $response ) || 200 != wp_remote_retrieve_response_code( $response ) ) {
-				@unlink( $browscap_zip );
-				return array( 1, __( 'There was an error downloading the Browscap data file from our server. Please try again later.', 'wp-slimstat' ) );
-			}
-
+			
 			if ( !file_exists( $browscap_zip ) ) {
 				return array( 2, __( 'There was an error saving the Browscap data file on your server. Please check your folder permissions.', 'wp-slimstat' ) );
 			}
 
+			if ( is_wp_error( $response ) || 200 != wp_remote_retrieve_response_code( $response ) ) {
+				@unlink( $browscap_zip );
+				return array( 3, __( 'There was an error downloading the Browscap data file from our server. Please try again later.', 'wp-slimstat' ) );
+			}
+
 			// Init the filesystem API
-			WP_Filesystem();
+			require_once( ABSPATH . '/wp-admin/includes/file.php' );
+			if ( !function_exists( 'wp_filesystem' ) ) {
+				@unlink( $browscap_zip );
+				return array( 4, __( 'Could not initialize the WP Filesystem API. Please check your folder permissions and PHP configuration.', 'wp-slimstat' ) );
+			}
+
+			wp_filesystem();
 
 			// Delete the existing folder, if there
 			$GLOBALS[ 'wp_filesystem' ]->rmdir( dirname( self::$browscap_autoload_path ) . '/', true );
@@ -103,10 +112,13 @@ class slim_browser {
 			$unzip_done = unzip_file( $browscap_zip, wp_slimstat::$upload_dir );
 
 			if ( !$unzip_done || !file_exists( self::$browscap_autoload_path ) ) {
-				return array( 3, __( 'There was an error uncompressing the Browscap data file on your server. Please check your folder permissions and PHP configuration.', 'wp-slimstat' ) );
+				@unlink( $browscap_zip );
+				return array( 5, __( 'There was an error uncompressing the Browscap data file on your server. Please check your folder permissions and PHP configuration.', 'wp-slimstat' ) );
 			}
 
-			@unlink( $browscap_zip );
+			if ( file_exists( $browscap_zip ) ) {
+				@unlink( $browscap_zip );
+			}
 		}
 
 		return array( 0, __( 'The Browscap data file has been installed on your server.', 'wp-slimstat' ) );
